Moved cli integration tests out of topic controller
