+++ tests/integration/test_topic_controller.py
@@ -1,12 +1,8 @@
 import json
-from typing import Dict, Any
 
 import confluent_kafka
 import pytest
-import yaml
-from click.testing import CliRunner
 
-from esque.cli.commands import apply
 from esque.errors import KafkaException
 from esque.topic import Topic, TopicDiff
 from esque.topic_controller import TopicController
@@ -57,19 +53,6 @@ def test_alter_topic_config_works(topic_controller: TopicController, topic_id: s
 
     final_config = after_changes_applied_topic.config
     assert final_config.get("cleanup.policy") == "compact"
-
-
-@pytest.mark.integration
-def test_topic_deletion_works(
-    topic_controller: TopicController, confluent_admin_client: confluent_kafka.admin.AdminClient, topic: str
-):
-    topics = confluent_admin_client.list_topics(timeout=5).topics.keys()
-    assert topic in topics
-    topic_controller.delete_topic(Topic(topic))
-    # Invalidate cache
-    confluent_admin_client.poll(timeout=1)
-    topics = confluent_admin_client.list_topics(timeout=5).topics.keys()
-    assert topic not in topics
 
 
 @pytest.mark.integration
@@ -133,115 +116,3 @@ def test_topic_diff(topic_controller: TopicController, topic_id: str):
     topic = Topic.from_dict(conf)
     diff = TopicDiff().set_diff("replication_factor", 1, 3)
     assert get_diff(topic) == diff, "Should have a diff on replication_factor"
-
-
-@pytest.mark.integration
-def test_apply(topic_controller: TopicController, topic_id: str):
-    runner = CliRunner()
-    topic_name = f"apply_{topic_id}"
-    topic_1 = {
-        "name": topic_name + "_1",
-        "replication_factor": 1,
-        "num_partitions": 50,
-        "config": {"cleanup.policy": "compact"},
-    }
-    topic_2 = {
-        "name": topic_name + "_2",
-        "replication_factor": 1,
-        "num_partitions": 5,
-        "config": {"cleanup.policy": "delete", "delete.retention.ms": 50000},
-    }
-    apply_conf = {"topics": [topic_1]}
-
-    # 1: topic creation
-    path = save_yaml(topic_id, apply_conf)
-    result = runner.invoke(apply, ["-f", path], input="Y\n")
-    assert (
-        result.exit_code == 0 and "Successfully applied changes" in result.output
-    ), f"Calling apply failed, error: {result.output}"
-
-    # 2: change cleanup policy to delete
-    topic_1["config"]["cleanup.policy"] = "delete"
-    path = save_yaml(topic_id, apply_conf)
-    result = runner.invoke(apply, ["-f", path], input="Y\n")
-    assert (
-        result.exit_code == 0 and "Successfully applied changes" in result.output
-    ), f"Calling apply failed, error: {result.output}"
-
-    # 3: add another topic and change the first one again
-    apply_conf["topics"].append(topic_2)
-    topic_1["config"]["cleanup.policy"] = "compact"
-    path = save_yaml(topic_id, apply_conf)
-    result = runner.invoke(apply, ["-f", path], input="Y\n")
-    assert (
-        result.exit_code == 0 and "Successfully applied changes" in result.output
-    ), f"Calling apply failed, error: {result.output}"
-
-    # 4: no changes
-    result = runner.invoke(apply, ["-f", path])
-    assert (
-        result.exit_code == 0 and "No changes detected, aborting" in result.output
-    ), f"Calling apply failed, error: {result.output}"
-
-    # 5: change partitions - this attempt should be cancelled
-    topic_1["num_partitions"] = 3
-    topic_1["config"]["cleanup.policy"] = "delete"
-    path = save_yaml(topic_id, apply_conf)
-    result = runner.invoke(apply, ["-f", path], input="Y\n")
-    assert (
-        result.exit_code == 0 and "to `replication_factor` and `num_partitions`" in result.output
-    ), f"Calling apply failed, error: {result.output}"
-    # reset config to the old settings again
-    topic_1["num_partitions"] = 50
-    topic_1["config"]["cleanup.policy"] = "compact"
-
-    # final: check results in the cluster to make sure they match
-    for topic_conf in apply_conf["topics"]:
-        topic_from_conf = Topic.from_dict(topic_conf)
-        assert not topic_controller.diff_with_cluster(
-            topic_from_conf
-        ).has_changes, f"Topic configs don't match, diff is {topic_controller.diff_with_cluster(topic_from_conf)}"
-
-
-@pytest.mark.integration
-def test_apply_duplicate_names(topic_controller: TopicController, topic_id: str):
-    runner = CliRunner()
-    topic_name = f"apply_{topic_id}"
-    topic_1 = {
-        "name": topic_name,
-        "replication_factor": 1,
-        "num_partitions": 50,
-        "config": {"cleanup.policy": "compact"},
-    }
-    apply_conf = {"topics": [topic_1, topic_1]}
-
-    # having the same topic name twice in apply should raise an ValueError
-    path = save_yaml(topic_id, apply_conf)
-    result = runner.invoke(apply, ["-f", path], input="Y\n")
-    assert result.exit_code != 0 and isinstance(result.exception, ValueError), f"Calling apply should have failed"
-
-
-@pytest.mark.integration
-def test_apply_invalid_replicas(topic_controller: TopicController, topic_id: str):
-    runner = CliRunner()
-    topic_name = f"apply_{topic_id}"
-    topic_1 = {
-        "name": topic_name,
-        "replication_factor": 100,
-        "num_partitions": 50,
-        "config": {"cleanup.policy": "compact"},
-    }
-    apply_conf = {"topics": [topic_1]}
-
-    # having the same topic name twice in apply should raise an ValueError
-    path = save_yaml(topic_id, apply_conf)
-    result = runner.invoke(apply, ["-f", path], input="Y\n")
-    assert result.exit_code != 0 and isinstance(result.exception, KafkaException), f"Calling apply should have failed"
-
-
-def save_yaml(fname: str, data: Dict[str, Any]) -> str:
-    # this path name is in the gitignore so the temp files are not committed
-    path = f"tests/test_samples/{fname}_apply.yaml"
-    with open(path, "w") as outfile:
-        yaml.dump(data, outfile, default_flow_style=False)
-    return path
